requirements.txt
langchain>=0.0.273
python-dotenv>=1.0
unstructured>=0.10
InstructorEmbedding>=1.0
torch>=2.0
sentence-transformers>=2.2
chromadb>=0.4
elasticsearch